Make foreman replace ${::facts} for puppet
The inclusion of ${ and } in a string make Puppet perform substitutions.
When importing, the smart-proxy must not perform such substitution,
it therefore should be done upon exporting the YAML ENC output.
Unfortunately, Puppet does not perform it upon reception of the YAML
document, hence we have to do it inside Foreman.