Host/hostgroup edit: hide internal Puppet classes by default
If a Puppet class is declared as "internal", it won't be displayed in
the "Puppet Classes" tab when editing a host or a hostgroup for
instance.

To declare a class as internal, one must add a "modeline" in the class
documentation itself. Here's an example:

    # Class: mymodule::myclass
    #
    # foreman: { internal-class }
    class mymodule::myclass {
        ..
    }

A modeline is a one-line JSON hash, preceded by "foreman:". Only
whitespaces are allowed before the prefix and after the structure.
"internal-class" must appear as a key, with whatever one wants as the
associated value.

In the "Puppet Classes" tab, one can display internal classes in the
"Available Classes" list by clicking on a new button, located on the
right of the "Filter classes" text field.

In the "Included classes" and "Available Classes", internal classes are
displayed in a light grey to be able to distinguish them from the normal
classes.