Settings with enum type
When declaring a setting, you can pass meta data,
which won't be stored in the database, but will be available in the instance.
It is an in memory map based on the setting name.

These are used to store the Array or Hash an enum typed setting must resolve to.
For Hashes, the raw_value will map to a key, and value to the associated value.
An Array is treaded as a Hash with both keys and values being identical.
Note that raw_default must map to a key too.

The values and default values are enforced to be mentionned in the enum.
The raw_ variants of default and value are used to retrieve the key of the enum,
rather than the associated value.

Using Hash, one can store an arbitrary object in a Setting,
providing a simple string key to control its value.

The optional fourth option to Foreman::DefaultSettings::Loader.set
is the meta data hash, whose key :enum is actually used
to associate the valid values of an enum typed setting.

I prefixed the initializers
because config/initializers/puppet.rb was required
for its requirement on 'puppet/rails',
before config/initializers/foreman.rb gets executed
because lib/foreman/default_settings/loader.rb now references
Host (app/models/host.rb) which inherits Puppet::Rails::Host...
Anyway, these are better prefixed!

I left config/initializers/inflections.rb as a symlink to its prefixed sibling,
because vendor/**/has_many_polymorphs*/lib/has_many_polymorphs/autoload.rb
requires "#{Rails.root}/config/initializers/inflections".